Add return type annotations to functions

+++ src/Utils.jl
@@ -2,7 +2,7 @@ module Utils
 
 using ITensors
 
-function bipartite_entropy(psi::MPS, seperator_index::Int)
+function bipartite_entropy(psi::MPS, seperator_index::Int)::Float64
     orthogonalize!(psi, seperator_index)
     _, S, _ = svd(
         psi[seperator_index],
@@ -18,7 +18,7 @@ function bipartite_entropy(psi::MPS, seperator_index::Int)
     real(SvN)
 end
 
-function center_bipartite_entropy(psi::MPS)
+function center_bipartite_entropy(psi::MPS)::Float64
     center = trunc(Int, length(psi) / 2)
     bipartite_entropy(psi, center)
 end
